requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
pydantic==2.9.2
python-jose==3.3.0
python-dotenv==1.0.1
alembic==1.13.3
email-validator==2.2.0
passlib>=1.7.4
bcrypt>=4.0.1,<4.1
python-multipart==0.0.9
pyserial>=3.5

